Replace broken dbus calls with new ones using filters
This breaks the dbus API from 0.2 because the args have changed.
However, this is necessary because the internal code that the 0.2 dbus
API calls used is no longer present, and these calls were broken.

The new API calls use the new filters system.  This is a much more
powerful way to construct "searches" against the internal tasks, and
provides some of the same functionality as had been present in the old
API calls, except that searching by tags is not yet available.  It looks
like maybe some infrastructure work is needed to support constructing
filters on the fly in order to provide tag querying, so that is still
TBD.