Hide notebook from the start
When gtg is initially started for the first time, there are no settings
to indicate whether or not the accessory_notebook should be shown, so it
falls back to the glade settings, which was set to visible.  But since
the closed task pane was not being enabled, all the user saw was an ugly
grey bar.  (LP: #583801)