Clean up some more messy debug messages
This leaves a one line message to display to the user, the rest is
displayed only if --debug is specified.  The information it provides is
really only interesting for debug purposes (and probably of limited use
even there), so this reduces the spewage to stdout when these errors are
present.