Fix for Bug #1265979 - Error when opening Preferences dialog
Even though having 'gsettings' not installed is rare, if it's not installed the preferences dialog failed to open up.
Now on opening the dialog, gsettings presence will be checked. If not found, the dialog will still open. Only the 'New Task Shortcut' feature will be disabled.
A label will be shown below the disabled option asking the user to install 'gsettings' first.