Fixed task urgency (LP #1258451).
(Not sure how this fix got missed. The Launchpad bug said it was already
committed.)

The calculated "urgent" due date for a task, i.e. the quantity used
for sorting the "Due" column, is wrong for a task with closed
sub-tasks.

Take these tasks:

* Task1 due:2013-12-17
   \
    '--Subtask1 due:2013-12-05 closed:2013-12-01

* Task2 due:2013-12-07

When the task browser shows these tasks, sorted as "Due ^", Task1 is
shown at the top (with due description "in 11 days") before Task2
(with due description "tomorrow").

This is because get_urgent_date() for Task1 is returning 2013-12-05,
even though that task has already been closed.

The fix is to ignore closed sub-tasks.