Fix infinite loop when openining up an editor
In the newer Gtk (tested on 3.14 that is included in Ubuntu 15.04), there was a slight change in semantics of forward_word_end. If there is a trailing white space, forward_word_end doesn't move to the end of buffer as it is not a word per se,  but returns an error (value False).

The following script runs in infinite loop under new Gtk:

```python

from gi.repository import Gtk

buf = Gtk.TextBuffer()
buf.set_text("word ")
it, end = buf.get_bounds()

while (it.get_offset() < end.get_offset()) and (it.get_char() != '\0'):
    it.forward_word_end()
    prev = it.copy()
    prev.backward_word_start()
    print('word: "{}"'.format(buf.get_text(prev, it, True)))
```

Fix is to check the return value of it.forward_word_end() in while loop:

```python
while it.forward_word_end():
    prev = it.copy()
    prev.backward_word_start()
    print('word: "{}"'.format(buf.get_text(prev, it, True)))
```

Fixes #90