Refactor translations, directories & CoreConfig
All of that functionality was hidden in GTG/__init__.py and
GTG/core/__init__.py that it makes hard to find it for newcomers.
Provided features are non-trivial and should be in separate files:

  - translations are now in GTG.core.translations
  - CoreConfig is now in GTG.core.config
  - directories are now in GTG.core.dirs

CoreConfig was breaking single responsibility principle as it was
providing configuration, constants for directories and various other
constants. CoreConfig is now clean:

   - it handles only configuration
   - all information about directories is stored in single place
     (GTG.core.dirs)
   - TAG constants were moved to GTG.core.tag