Handle changing tasks/deleting tasks via liblarch
Browser and editor were asking manager to change status of a task/delete
a task. That approach doesn't scale as there are other ways how to
change that (e.g. via DBUS). Proper solution is to subscribe to liblarch
callbacks for everytime when a task is modified or deleted.

Result is manager object that has one less responsibility and thus
resulting in cleaner architecture.