Make anonymize_task_file script independent of GTG
Import from GTG can lead into problems with import paths, be dependent at the latest installed version of GTG, etc. In other words, it's not worth the hassle.
It's revert of 44fb77d58ded959e9ebcb5f60bb6a329ebf5e515 for this file.

Also handle the case when task file is listed as relative to the project file.