Resolve the deprecated widget properties
Some of the widgets within GTG used outdated, deprecated properties which are not a standard in GTK3+ anymore.
This PR therefore solves this: removes some of the unnecessary items and replaces others with the most
up-to-date properties and values.

This commit is a response to #113. I used Fedora 22 which has the latest standards implemented and thus raises errors when
some deprecated item is used.

Fixes #113