Refactor preferences and  the start of the day functionality
Start of the day will from now on be set in a single GtkEntry rather than two. This means that the time parser has to be changed, and some further possibilities of time input will be added.

I am implementing new time_parser function which uses regular expressions to analyse the user input and yield the correct time format consequently.

The entire code is simplified now beacuse preferences.ui acts only like a crossroad for other individual parts of GTG preferences which will be since now implemented in separate modules.