Implement search to a separate search entry and modify quickadd
This commit aims to clarify the distinction between search and quickadd. Search
is now accessible from the headerbar and offers immediate, automatic searching
after each key stroke. This made the search more conveninet, however the
functionality is not very fluent or fast yet.

Search tags are now not saved automatically, but only after the user decides to
do so using the "Save search" button. Such a search tag can then be, as before,
re-accessed from the tag sidebar.

As for the quickadd, it has been moved to the bottom of the GTG window and
features only adding and opening of tasks.

Fixes #137, fixes #140, fixes #143,