Remove task count from title in task browser
This commit is a response to #117. The feature of displaying the number of tasks in stock is not needed anymore as we transfered to header bar and we are not interested in providing this kind of information there anymore.

Therefore, I removed the parts of code which took care of displaying this information.