Drop gtcli and gtg_new_task
commit 2d380bb2f856e removed the dbus service needed by gtcli and
gtg_new_task, rendering them non-functional. In the future the same
functionality may re-appear as arguments to the main application
that passes it on to the running main instance, or via a new dbus
service in the main GtkApplication claimed dbus name. Delete the
broken utilities meanwhile, along with various references to them,
now that there's no setup.py that tries to install them (they were
never hooked up to the meson build).

Closes #310