Restore desktop file GenericName translation for Slovak and Spanish
commit c83b00711b changed desktop file to be translatable via PO files,
and thus had to remove already translated GenericName and Comment for
sk and es languages. Restore the GenericName translation for them via
PO file. Comment is intentionally not restored, because a fuzzy entry
was already generated for them, and the previous translation in
desktop file isn't an exact translation from English.