docs: Drop Russian translation
The existing old translation was done as if a fully separate user
manual, not as a translation of English, at least as far as long-term
maintainability is concerned. The new method, as hooked up in meson
port, does this in the modern standard way, where the translation
ends up in a gettext catalog and the build/install process will create
the translated *.page files.
Preserving the translation is of questionable value, as the English
text has, and will, change substantially and it may be easier to just
start the translation from scratch. However one can always refer to
the old translation if needed.
Remote the old unused translation files, to make room for any new-style
translation of the user manual to Russian.