Rework translation handling and make it work with non-default prefixes
.. and when ran via gtg.sh

One test insides tests/tools/test_dates.py now tries to test translations
without a full gettext setup due to the move of this to the main gtg
binary. However tests shouldn't rely on the running environments
locale anyhow, so whatever it is doing sounds better without honoring
the locale than honoring it as before. If this test is needed, it should
set up its own specific language environment (or iterate through available
translations) and test that way (and automatically skip for locales that
aren't available on the running system).

Fixes #304