meson: Explicitly install python sources to purelib path
Despite the documentation stating that it's done by default, it wasn't
doing that until very recently. However it was in the case of
get_install_dir, so they were inconsistent and breaking gtg.sh usage
on systems where purelib and platlib differ for the used python.
Request pure explicitly to workaround the meson issue that was fixed
by https://github.com/mesonbuild/meson/pull/6848 in meson.