Update introduction topics for clarity
The gtg-quickstart page was duplicative of what is on the
index and provided in the Overview of Main UI.
This page was removed and meson.build was updated.
Updates were also made to the index to consolidate
the Overview of the Main UI and the Keyboard Shortcuts
into an Introduction section.