Add translation context for the tasks view
The open tasks view (to view all open tasks) is rightfully called
"Open", however, it is also used as an verb (like in the Export plugin).
This is problematic for languages that has distinct words for those two
situations (such as german: "Offen" for the former one, and "Öffnen" for
the latter one). Because the latter is the more "usual" case, the name
of the view names get an special gettext message context.