Export plugin: Remove useless settings dialog
There was an settings dialog to set whenever it should include the menu
button to export be visible or not.
However, it is basically useless, since if an user would hide it, it
then can't interact with the plugin anymore. It is basically
deactivating the entire plugin.
The more "invisible" setting of the last template used should be
continued to be stored and retrieved as usual.