Mark .gtg-plugin files translatable
The way this is done isn't something I really liked to do.

The probably biggest hack would be introducing .gtg-plugin.desktop
files, which are just used so that xgettext (which is run by ninja
gtg-pot) uses the Desktop-language on them, so the appear in the .pot
and .po files.

A current issue is that when translation is being updated, a re-launch
(via launch.sh) doesn't regenerate the files. To force regenerating the
files, run rm -f .local_build/GTG/plugins/*.gtg-plugin