Context Menu
Adding GUI entries/actions.

Further information:
	-In the task editor, two buttons were added to
	Enable/Disable and pick a recurring term.
	-In the main_window, actions were added to
	quickly select a recurring term.

Activated contextmenu options for repeating tasks.

FIX: date of repeating subtasks not set properly.

When a subtask of a "normal" task is duplicated, the due date would not
be set correctly due to the method add_child which set a subtask's date
to its parent's. That is why add_child will no longer set the dates of recurring tasks.