Chore: add logging when duplicating a task
Chore: Centralized way to inherit recurrency

While a task inherited its recurrency in the on_add_subtask method in
main_window, a better way is to call the fuction in add_child.

Fix bug Centralized way to inherit recurrency

In certain cases, set_parent is used to add a subtask. For instance,
when a subtask is "dropped" through the TreeView, inherit_recursion
was never called.