Use GLibs cli parser and use application.open
The primarily goal was to use the GApplication command line parser, so
that an user can also access gtk and GLibs command line arguments.

However, the opening tasks part made this much bigger, since it is done
from the activate-signal, which is usually reserved to open without any
files. With the new system, the open-signal is used to open files, so it
had to be ported to over there.

Initializing the browser window wasn't possible in the startup-signal
(it crashes, and it seems you aren't supposed to do that), so this is
done by both activate- and open-signals, in shared_init.

An weird oddity I've found is that in "the" documentation for python,
the do_open just accepts 2 arguments: files and hint. However, I was
getting argument mismatch errors, and I've noticed that I got a number
as the second parameter, and by looking at the C docs, the second
parameter is the number of files, which matches what is passed in.
Since files is a python list, this isn't needed, but checked anyway,
just in case.
https://lazka.github.io/pgi-docs/Gio-2.0/classes/Application.html#Gio.Application.do_open
    do_open(files, hint)
https://developer.gnome.org/gio/2.66/GApplication.html#GApplication-open
    void
    user_function (GApplication *application,
		   gpointer      files,
		   gint          n_files,
		   gchar        *hint,
		   gpointer      user_data)

The files in the open-signal are GFiles, which seem to change the URI a
tiny bit by adding a final slash (gtg://a → gtg://a/). Because of this,
the old code wouldn't find the task, so I went ahead and parsed the URI
completely using urllib.parse.
GLib also has URI parsing, but it seems they aren't exposed to python.

Also, GLib also translates the help strings if they are translated.
However, I needed to put GTG/gtg.in manually in po/POTFILES.in, but even
then, xgettext defaulted to using C instead of python (rightfully).
An workaround would be creating an symlink gtg.py.in and inputting that
to xgettext, but I didn't wanted to deal more with this, so I didn't
bother to continue.