Use _("...").format instead of _(f"...")
The latter would format the string first before translating, and having
an translation for every possible combination is basically impossible.

I am using .format() specifically because diegogangl commented about
having to drop it for .format(), so I am using that.

I tried to not modify the original string to not break existing
translations, but I only was half-successful.