Make module-level translations work
I saw this being done in D-Feet:
https://gitlab.gnome.org/GNOME/d-feet/-/blob/b312c714fc01c3d14e2c6e1da6e87645674cca05/src/d-feet.in#L38-41

The problem is that some code uses gettext "at the module level", which
is run when the module is being imported.
Before this change, basically all of GTG gets imported, before locale
gets set up.
After this change, it gets setup before importing it, so it'll work as
expected. Not sure why I haven't thought of this earlier.

I've put this in a if __name__ == "__main__" block, because the original
code was also in such block.