Migrate Timer from dbus-python to Gio
Initially, it didn't work. The reason was that system_dbus object is
created on that line and when it goes out of scope, it gets destroyed
and the signals get disconnected.

I assumed that the connection would stay open as part of Gio and not be
created on first use (althrough it makes sense when implementing it).
Also, I assumed that the signal would make it stay connected.

This took days to figure out, I even wrote a more "minimal" python
version, and even a C version which let me find out the issue.
See https://gitlab.gnome.org/-/snippets/1288

Thus the fix is to just keep a reference to it in the timer object
itself.