Move Application init to startup signal
Previously, a bit of stuff was done in the constructor, which (likely)
hurts being D-Bus activate-able, since the "remote" instance (aka when
trying to execute GTG again which then would hand off to the existing
instance) would re-initialize only to not use that stuff.

Also changed from using an list-comprehension [x for y in z] to an
"normal" for-loop, since it is two lines anyway and the result isn't
used.