Don't print errors for the same icon multiple times
Previously, when getting an icon error while rendering tasks (such as
when the user changes icon themes), it won't find the icon and prints an
error.

This is fine, but it spams the terminal/console/logs with essential
useless stuff, thus it is made sure only to print it once (per
renderer).