Fix bug when deleting subtask text
If the user deleted all text in a subtask all the way to the checkbox it
could get detected as the checkbox itself having been deleted. In that
case the subtask tag would be unmarked and it wouldn't be possible to
write a new title for the subtask.

We now make sure that the checkbox is there, so new text can be detected
and the subtask renamed.