Replace custom icons selector with emoji picker
"Custom icons" was kind of a misnomer, since you could only select from
a short list of icons. These icons probably changed over time since they
were introduced, and at the time of this commit had come to have
different styles and sizes. Most of them were useless as icons or
duplicates.

This commit starts replacing them with emojis, which are basically a
+3,000 icon set baked into the toolkit.

ref #85