Remove FilteredDatastore class
The purpose of this class was to simulate private methods in Python.
Backends would only have access to the filtered datastore and thus only
see a few methods. However, backends also have access to the requester
which has access to the regular datastore, so they can access all methods
just the same. Let's just lower complexity and remove this class.