Update recurring calendar to today or selected date
Previously it'll always default to the same year and month, by removing
the attributes it defaults to being today.

In code, it is then updated to match the current recurring status when
applicable.

The monthly selection is now always using January since it has 31 days,
the max days in a month.
To make sure this detail isn't seen when going to the yearly calendar,
this is made hidden with an ugly loop, trying out the next possible
month.