Workaround: Separate saved searches with tags name
Since unfortunately saved searches and tags share the same namespace, it
makes it difficult to create a saved search with the exact name as an
existing tag.

As a current workaround until we separate them, we just prefix the saved
searches with SEARCH_TAG_PREFIX, which an user is unlikely to use.
From what I can see, it still displays properly in the UI after some
fixes, aside from existing bugs.

In the XML file this prefix isn't added, because it would be redudant
and make it easier to deal with in future versions.