stop using GtkBox child properties in code
This is pack_start/pack_end. Also, I ignored the backends stuff because
it seems to never actually run

container_add+vexpand/hexpand will be a lot easier to port to the new
api.

Also, I am from the far future when I didn't replace the expand in
add_widget_to_taskeditor because that is stupid

Also, setting Fill but not Expand doesn't do anything! And that is
everywhere!