port plugin api to GMenu
This is broken now that we use Gmenu instead of arcane modelbuttons.

The maintained plugins have been ported and I can confirm that they work
correctly.

The exception is hamster which I didn't port yet since it uses them in
a non trivial way, modifying them (which is a lot more conveluted with
gmenu). Will do that after GTK4, and after I can figure out how to run
that plugin without dbus errors

sendEmail is a bit different, in this case I used an action group to
associate to the correct open editor, instead of a global action.
I used the style "editor" (component) "_plugin_", "email" - the name of
the plugin. Also, I have done this after I travelled back in time from
the very far feature where I realised that this is needed, and that you
need to not use '.' in the action group name or GTK will crash.