gtk4: adapt to focus changes
This is mainly removing the stupid can_focus>False that people keep
putting into their UI files. First, it has a different meaning in GTK4
(you usually dont need it). And you don't need it in 99% of cases in
GTK3, but Glade decided to throw it down anywhere, so many UI files are
full of this useless property

Default focus has changed, instead of the billion has/is/can focus
properties you set a property on the GtkWindow, this will be accoutned
for later when the main_window ui file will actually contain the window