gtk4: editor: use composite template
Gtk.Builder.connect_signals is gone, the replacement needs to be
implemented by the language binding, for PyGObject it isn't.

Composite Templates allow easy connecting of signals, to avoid
connect_signals we use composite templates

The part with the dict of lambdas got expanded quite a lot, maybe I
should have used a for loop that calls the Gtk.Template.Callback()
decorator on them? However I don't know if callbacks maybe need a
reference of them to exist to function