gtk4: port TagContextMenu to Popovers
quick fact: I originally used app.add_action and prefixed the action
names with "tags_popup.", and this was because using a actiongroup would
cause a crash. And guess what? I in the far future figured out why the
crash was happening, and also found a convenience function
install_action to simplify installing actions class-wide.