gtk4: port error handler
The fundamental issue was that positional arguments are actually gone
now (not only deprecated). DialogFlags is not used because it can only
be used when using the dialog constructor, which seems to be broken with
messagedialog in pygobject

I set the scrolledwindow min-height because something with the margins
of messagedialog changed in GTK4, making the default way too small.

title -> text, because text looks like a title now, and title doesn't
really. This means that for non-title we only have secondary-text, while
we previously used 2 descriptions, so I cated them with +2*"\n", this
loses the justification though