gtk4: adapt to about_dialog changes
It is not a GtkDialog anymore (and stop using those type hints on it
too).

This has the consequence of not being able to add custom widgets for the
dependencies/status anymore.

I have instead used the new System Information tab which I think is the
closest property there is to what we need.
I have also reworded it quite a bit, and made it no longer use markup as
it isn't supported.

Also, now it seems you can't change the about dialog info after the
fact, so now we instead construct one the moment you click the button
for that specific plugin. I kept the old look. (Even the fact that it is
not transient+modal on the pluginsdialog which I think is a lot better)