README: Install python3-gi-cairo for debian
This is required for getting GTG to work, otherwhise the following error
would occur:

TypeError: Couldn't find foreign struct converter for 'cairo.Context'

It seems that package is installed by gedit, which is a default
applictaion for GNOME, so during package testing it might be easy to
miss.

Since python3-gi-cairo depends on python3-cairo, I "replaced" that
dependency, so the line isn't getting too long.