gtk4: don't use get_template_child (plugin api)
get_template_child isn't the proper solution to Gtk.Builder.get_object
as it can only give things that are already defined in the template.

The main use of this was for the plugin api.

This might be a bit unideal, but now instead the task editor and main
window have public methods to get the widgets get_tempalte_child was
used for, and the plugin api calls this instead.