gtk4: port deletion confirmation dialogs
run() is blocking, and removed. Use response signal instead.

This is a bit complicated by the fact that these dialog use functions
also return something, so they are changed to _async( that take a
callback argument, similar to how many GTK apis work (except without
that _finish and asyncresult garbage)

Previously these also used regular text, but bold for what was going to
be done. However in GTK4 GtkMessageDialog:text is the title, this is a
problem as secondary-text (which is the small extra details text) is
already used. So instead the size of the title is reduced