gtk4: bringup TaskEditor
Make taskeditor mostly work with GTK4.

Notable things to do:
	* Make font selection use css instead of legacy override methods
	* Rethink the recurring menu (fundamentally broken)

Popovers now have a built-in margin, the margins of all the boxes inside
have been reduced by 6px which seems mostly correct.

Setting and getting the position of the window is an archaic X11 concept
not supported in GTK4, that is removed.

the insert_action_group('app', app) and self.insert_action_group('win',
app.browser) don't make sense to me, and those lines broke actions, and
caused cryptic GObject weakref errors. They are simply removed

The popovers are now shown/hidden on window focus out/in as they are now
real windows (when entry is focused). This allows for them to not float
during alt-tab on X11. Note that if the focus is **inside**  the popover
when this happens, then when you return they are still popped down, and
you have to focus the entry twice next time (eventcontroller doesn't
fire for some weird reason), I didn't quite figure that part out.

GtkTextTag::event is gone, and you cannot simply replace it with an
eventcontroller as they are not real widgets.
Instead now the textview itself has a click gesature, and it finds the
tag at the clicked point and calls a function on the tag with convenient
arguments related to clicks.

A bunch of get_iter_at functions have been changed to return a tuple,
instead of a single value. The second value in the tuple is all that is
needed for this usecase.