gtk4: rework recurring_menu
Now uses a menu model and actions instead of manually constructed
widgets.

I first wanted to use a stateful action for toggling, but I am using a
propertyaction for the checkbutton, and later I realized this actually
seems like quite a good way of doing it.

Now that the RecurringMenu is itself the widget, all the callbacks are
also in that file, which imo makes more sense and is a lot easier to do.