flatpak: use master SDK
We need a feature only available in GTK 4.5/4.6 for properly porting
cell_renderer_tags.

The stable 41 runtime only offers GTK 4.4

Note that flathub doesn't have the master sdk.
An alternative possible solution would be to bundle GTK instead (However
for some reason the app icon in the aboutdialog breaks then)

dbus-python needs to be updated to work with the updated runtime.