gamify_plugin: port to GTK4
Main change is preferences window is a separate composite template
class (for the signal autoconnect). It does maybe seem slightly weirder
like this.

I tried to preserve the old extremely broken design.

This also used get_children() and storing data in widgets, and getting
data from widgets by traversing their widget hierarchy, which is very
discouraged in GTK4. widget.get_children() -> list(widget) (this is a
pygobject override, the other method requires a ton of calls for
iterating)

There is a warning when disbaling the signal handler for requester
status-changed (with by_func too)