gtk4: tweak treeview paddings
To look like GTK3, note that this is an integer, and the exact correct
amount for both treeviews is a pixel off.

I first actually was setting the ypad on the tagcellrenderer in the task
treeviews, however then a lot later I noticed this actually made the
size inconsistent with tags that don't have any tags. So I switched to
setting it on the text instead, which is consistent and probably better
anyways due to how font size may change.