export_plugin: port to GTK4
I didn't figure out the dependencies fully, so I only tested the basic
export actions (although the export actions don't use GTK so it should
be fine). I can confirm they work correctly.

The filechoosernative bug I also encountered while porting another app.
I tried a possibly better workaround of disconnecting the signal handler
on first run however that didn't help.

GtkImage changed to GtkPicture because in GtkImage the pixbuf doesn't
get scaled up fully to take up all the available space anymore.

I opted to not port that to a composite template and instead just
connect signals manually.
For most of these plugins porting to composite templates would be very
weird because the UI is phylosophically "part" of the plugin object itself, but the
plugin object is not a GtkWidget.

The margins at the bottom of the dialog changed by 3 px as I made the
layout a bit more sane (Toplevel box has the margins now). And the fact
that they were inconsistent by 3px was probably an oversight anyway.