hamster_plugin: port to GTK4
The main change here is the migration to GMenu instead of widgets. As
the menu items are immutable, all functions that need to modify one need
a reference to the plugin api (which is different for each task).
So all functions that accept menu items now also need you to provide the
corresponding plugin_api. I also thought about maybe putting the plugin apis
in the menu_items dict.

Also, the prefs.ui file was like 90% useless properties (however
has-entry doesn't make sense there, so I removed it). I didn't replace
builder.connect_signals with a template as there was only one signal.