urgency_color_plugin: port to GTK4
GtkFrame labels are now below the frame, not in it, so it looks a bit
weird. (and not using a custom label widget doesn't change that,
although that wouldn't be possible anyway as we use markup (replace with
css?))

I tried a slightly different approach to compositing preferences this
time, now preferences stores the actual pref dict, and the plugin "gets"
a reference to it.