untouched_tasks: port to GTK4
This time I tried to experiment with a different way of approaching
porting preferences.

Since the prefs look very much like a GtkDialog, and
builder.connect_signals was used only for functionality that is provided
by GtkDialog already, I have ported the window to use a GtkDialog.

I don't understand from where the space between the sentence box and the
vbox2 was comming from in the original (there were no margins or child
spacing set!). So I have simply set an appropriate margin of 20px on the
box.

The buttons look a bit different as I would need to access the box
inside the action_area to make them align to the left.

Also, you do realise using a box with widgets and labels like that doesn't
work with localization, right?

Also, yay: all plugins are ported!