gtk4: stop adding custom GtkImage children to buttons
In GTK4 GtkButton has an icon-name property that does this for us. It
also applies the image-button style class which is required for proper
functioning in Libadwaita (without it, it gets treated like a button
with custom content that cannot become flat in toolbars)

Also note that GtkMenuButtons didn't support custom children until very
recently in GTK4, so without this they would require a higher minimum
version.