gtk4: startdate_calendar: add "Today" button
In GTK4 selecting the currently selected date in a GtkCalendar does not
change the day property or emit the day-selected signal anymore, with no
way to change this behavior.

To still allow selecting the current date as the start date easily, add
a today button.

I don't think this is necessary for the duedade calendar as there is a
"now" button, and how often do you create tasks that are due today
anyway?