Revert meson plugin translation apply thing
Since meson 0.60, the `build_always_stale` option was removed:
https://github.com/mesonbuild/meson/commit/6b1a80024c79876dc3b75f2b525be24ca8fe671a

However, meson doesn't make such files depend on the translation .po
files, which is a known bug:
https://github.com/mesonbuild/meson/issues/2621

So with that option taken away, I tried passing a list to the input
which includes all .po files, since I saw it only uses the first element
for the translations, but that fails when trying to generate po files.

Since it seems you can't dynamically add dependencies after the fact,
this seems like a dead end, so I am reverting this for now to make it
compile on meson 0.60.