Adaptive tomorrow button to make main window smaller
A bug report was that the minimum width of the main window is a bit too
high for some resolutions (when trying to use GTG half of it). A
proposed solution would be to use an icon, but it is rather hard to find
a good system-provided one.

What I've thought is that maybe just display the "Start tomorrow" text
while the window is big enough, and only the icon when it gets smaller.

I couldn't find a good way to do that with only using GTK stuff, so I
whipped my own (and first) widget that displays an widget depending on
the (minimum) size of its children widgets, which then would be the
"Start tomorrow" label, and an Image for the icon.

This widget is rather tailored for GTG in that use case only, which means
not everything has been implemented yet, like different size request
modes. Maybe this'll be done later.
Another thing to look out is easy GTK 4 migration.

Additionally, since the label is a bit long compared to the icon and
"snapping" doesn't work (just leaves large space in the button, and I
couldn't figure out how to "fix" this), I made the label be ellipseable
and set a minimum width to "overshadow" the icon.

Also, be careful when doing changes, I've encountered accident
exceptions-loops that caused my window manager (i3wm) to "freeze".
However, I could just switch to an virtual terminal, log in, kill GTG,
and restart i3wm from there (via i3-msg, very nice).