Use bgcolor to blend not white
Previously it would always blend on white background when in RGB mode,
but this isn't really what we want, but rather the actual background
color. Luckly, we have the `bgcolor` parameter we can use.

Additionally, the alpha (brightness) has changed to the old version,
otherwise the color blending will produce wrong colors because
`brightness > target_brightness`, so alpha can be over 1 (the maximum).