Don't loop forever with invalid quick add date
Quick entering "Do a thing due:meh" causes GTG to completely freeze, and
need to be killed. This fixes it to treat tokens with invalid dates the
as other invalid tokens, ignoring them (and removing them from the
input text).

This issue was caused since the the `continue` in the parsing loop
didn't remove the token, causing it to loop forever on the invalid
token.