gtk4: port backends dialog
I had ignored it in the past, but now it is actually used, so now it is
ported.

I made a small UI change for the main sidebar, as .inline-toolbar has
been removed, and reinventing that ancient pattern would have definetely been
more difficult.

You may also notice some minor changes in the look, as all the
previously used, now removed GtkAlignments used very stupid random
margins instead of using a sane box/spacing layout. So I simply redid
that, and it is a lot simpler to port, and immediately looks a bit nicer.

CellRendererPixbufs now use icon-name instead of that pixbuf from icon
name func, as porting that approach directly would be quite stupid.
Especially as GtkIconTheme gives GtkIconPaintable, and
CellRendererPixbuf needs a GdkTexture. And to convert a GtkIconPaintable
to a GdkTexture efficiently you need to do some manual GskRenderNode and GskRenderer
stuff that is broken in PyGObject.

Instead of getting a random insensitive GtkStyleContext color, now we
simply use the cellrenderer's sensitive property for disabled backends.