Fix adaptive widget offset when maximized in gnome
For some reason when using gnome-shell and maximizing GTG the widget
will be offset (just enough you can still see part of it outside the
button).

After reading more code it seems we need to resize the Gdk.Window when
doing size_allocation (which some of the GTK code doesn't do because it
was hidden behind some CSS helper functions it seems). We just call the
Gtk.Widget.do_size_allocation because they do it already for us.